+++ app/Http/Controllers/RelationshipsController.php
@@ -8,8 +8,7 @@ use Illuminate\Support\Facades\Redirect;
 
 class RelationshipsController extends Controller
 {
-    public function __construct()
-    {
+    public function __construct() {
         $this->middleware('auth');
     }
     /**
@@ -17,8 +16,8 @@ class RelationshipsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
-    {
+    public function index() {
+        $this->authorize('show-relationship');
         $relationships = \montserrat\Relationship::paginate(100);
         return view('relationships.index',compact('relationships'));   //
     
@@ -29,9 +28,9 @@ class RelationshipsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
-    {
-        //
+    public function create() {
+        $this->authorize('create-relationship');
+        
     }
 
     /**
@@ -40,9 +39,8 @@ class RelationshipsController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
-    {
-        //
+    public function store(Request $request) {
+        $this->authorize('create-relationship');
     }
 
     /**
@@ -51,9 +49,8 @@ class RelationshipsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
-    {
-        //
+    public function show($id) {
+        $this->authorize('show-relationship');
     }
 
     /**
@@ -62,9 +59,8 @@ class RelationshipsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
-    {
-        //
+    public function edit($id) {
+        $this->authorize('update-relationship');
     }
 
     /**
@@ -74,9 +70,8 @@ class RelationshipsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
-    {
-        //
+    public function update(Request $request, $id) {
+        $this->authorize('update-relationship');
     }
 
     /**
@@ -85,9 +80,9 @@ class RelationshipsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
-    {
+    public function destroy($id) {
+        $this->authorize('delete-relationship');
         \montserrat\Relationship::destroy($id);
         return Redirect::back();
-        }
+    }
 }
